+++ planner/core/casetest/physical_plan_test.go
@@ -2586,24 +2586,3 @@ func TestIndexMergeOrderPushDown(t *testing.T) {
 		tk.MustQuery("explain format = 'brief' " + ts).Check(testkit.Rows(output[i].Plan...))
 	}
 }
-
-// It's a case for index merge's order prop push down.
-func TestIssue43178(t *testing.T) {
-	store := testkit.CreateMockStore(t)
-	tk := testkit.NewTestKit(t, store)
-	tk.MustExec("use test")
-	tk.MustExec(`CREATE TABLE aa311c3c (
-		57fd8d09 year(4) DEFAULT '1913',
-		afbdd7c3 char(220) DEFAULT 'gakkl6occ0yd2jmhi2qxog8szibtcqwxyxmga3hp4ktszjplmg3rjvu8v6lgn9q6hva2lekhw6napjejbut6svsr8q2j8w8rc551e5vq',
-		43b06e99 date NOT NULL DEFAULT '3403-10-08',
-		b80b3746 tinyint(4) NOT NULL DEFAULT '34',
-		6302d8ac timestamp DEFAULT '2004-04-01 18:21:18',
-		PRIMARY KEY (43b06e99,b80b3746) /*T![clustered_index] CLUSTERED */,
-		KEY 3080c821 (57fd8d09,43b06e99,b80b3746),
-		KEY a9af33a4 (57fd8d09,b80b3746,43b06e99),
-		KEY 464b386e (b80b3746),
-		KEY 19dc3c2d (57fd8d09)
-	      ) ENGINE=InnoDB DEFAULT CHARSET=ascii COLLATE=ascii_bin COMMENT='320f8401'`)
-	// Should not panic
-	tk.MustExec("explain select  /*+ use_index_merge( `aa311c3c` ) */   `aa311c3c`.`43b06e99` as r0 , `aa311c3c`.`6302d8ac` as r1 from `aa311c3c` where IsNull( `aa311c3c`.`b80b3746` ) or not( `aa311c3c`.`57fd8d09` >= '2008' )   order by r0,r1 limit 95")
-}
