+++ planner/core/issuetest/BUILD.bazel
@@ -1,0 +1,11 @@
+load("@io_bazel_rules_go//go:def.bzl", "go_test")
+
+go_test(
+    name = "issuetest_test",
+    timeout = "moderate",
+    srcs = ["planner_issue_test.go"],
+    flaky = True,
+    race = "on",
+    shard_count = 50,
+    deps = ["//testkit"],
+)
